fix (migrate): merge `Homepage` to `index`, get rid of theme models because of no longer access `localStorage` in Nextjs
